+++ MyKnowledgeManager/src/MyKnowledgeManager.Core/Services/KnowledgeTagRelationService.cs
@@ -1,6 +1,5 @@
 ﻿using Ardalis.GuardClauses;
 using Ardalis.Result;
-using MyKnowledgeManager.Core.Entities;
 using MyKnowledgeManager.SharedKernel.Interfaces;
 
 namespace MyKnowledgeManager.Core.Services
@@ -33,14 +32,18 @@ namespace MyKnowledgeManager.Core.Services
             return knowledgeTagRelations.ToList();
         }
 
-        public async Task<Result<bool>> RemoveKnowledgeTagRelationAsync(string id)
+        public async Task<Result<bool>> RemoveKnowledgeTagRelationAsync(string id, string userId = null)
         {
             Guard.Against.NullOrEmpty(id, nameof(id));
+            
+            if (userId is null) return Result.Unauthorized();
 
             KnowledgeTagRelation knowledgeTagRelation = await _repository.GetByIdAsync(id);
 
             if (knowledgeTagRelation is not null)
             {
+                if (knowledgeTagRelation.UserId != userId) return Result.Forbidden();
+
                 try
                 {
                     await _repository.DeleteAsync(knowledgeTagRelation);
@@ -52,7 +55,7 @@ namespace MyKnowledgeManager.Core.Services
             }
             else
             {
-                return false;
+                return Result.NotFound();
             }
 
             return true;
@@ -67,9 +70,11 @@ namespace MyKnowledgeManager.Core.Services
             return knowledgeTagRelation;
         }
 
-        public async Task<Result<IEnumerable<KnowledgeTagRelation>>> GetKnowledgeTagRelationsAsync()
+        public async Task<Result<IEnumerable<KnowledgeTagRelation>>> GetKnowledgeTagRelationsAsync(string userId = null)
         {
-            return await _repository.ListAsync(new KnowledgeTagRelationsSpec());
+            if (userId is null) return Result.Unauthorized();
+
+            return await _repository.ListAsync(new KnowledgeTagRelationsSpec(userId));
         }
 
         public async Task<Result<KnowledgeTagRelation>> UpdateKnowledgeTagRelationAsync(KnowledgeTagRelation knowledgeTagRelation)
@@ -121,11 +126,13 @@ namespace MyKnowledgeManager.Core.Services
             return true;
         }
 
-        public async Task<Result<IEnumerable<KnowledgeTagRelation>>> GetKnowledgeTagRelationsByKnowledgeIdAsync(string knowledgeId)
+        public async Task<Result<IEnumerable<KnowledgeTagRelation>>> GetKnowledgeTagRelationsByKnowledgeIdAsync(string knowledgeId, string userId = null)
         {
             Guard.Against.NullOrEmpty(knowledgeId, nameof(knowledgeId));
 
-            return await _repository.ListAsync(new KnowledgeTagRelationsByKnowledgeIdSpec(knowledgeId));
+            if (userId is null) return Result.Unauthorized();
+
+            return await _repository.ListAsync(new KnowledgeTagRelationsByKnowledgeIdSpec(knowledgeId, userId));
         }
     }
 }
